component field parsing in entity template works + minor refactorings for component templates

+++ RoguelikeToolkit.Entities.Tests/Attributes.cs
@@ -2,7 +2,7 @@
 {
     public class AttributesComponent
     {
-        public int Strength;
-        public int Agility;
+        public double Strength;
+        public double Agility;
     }
 }
